+++ src/components/Theme/Theme.utilities.ts
@@ -1,2 +1,13 @@
-export const getRootThemeEl = () =>
-	document.querySelector("[data-rs-root]") || document.documentElement;
+export const getRootThemeEl = (scopeEl?: HTMLElement | null): HTMLElement => {
+	if (!scopeEl) return document.documentElement;
+
+	if (
+		scopeEl.hasAttribute("data-rs-root") ||
+		scopeEl === document.documentElement ||
+		!scopeEl.parentElement
+	) {
+		return scopeEl;
+	}
+
+	return getRootThemeEl(scopeEl.parentElement);
+};